name_parser: Address remaining codeclimate issues
The very low number of lines per file forces us to split up the stuff
into two files...
To still keep this a bit organized (and with the background that the
code propably will end up in the font-patcher file anyhow), all this
stuff is put into a dedicated subdir.

The call changes thus:

$ fontforge bin/scripts/name_parser/name_parser_test1 src/unpatched-fonts/**/*.[ot]tf 2>/dev/null
